+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Generate the declaration dynamically (#89)
 * Fixed bug concerning the `store-data` event
 * Fixed bug in Piral CLI regarding Pilet scaffolding
+* Added new optional library `piral-inferno` (#86)
 
 ## 0.8.3 (October 28, 2019)